requirements.txt
streamlit>=1.28.0
openai>=1.3.0
prettytable>=3.10.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.80.0
python-dotenv>=1.0.0
